Move AI data classes to data.model packages to prevent ProGuard obfuscation.
This commit identifies and moves data classes used for structured AI output
via GemmaClient to packages covered by ProGuard 'keep' rules.

Specifically:
- Moved CharacterIdeas to com.ilustris.sagai.features.newsaga.data.model.
- Moved OnboardingContent, OnboardingPage, and OnboardingConfig to
  com.ilustris.sagai.features.onboarding.data.model.
- Updated all import references in the codebase.
