+++ app/src/main/java/com/ilustris/sagai/features/onboarding/data/OnboardingModels.kt
@@ -1,21 +1,5 @@
 package com.ilustris.sagai.features.onboarding.data
 
-data class OnboardingConfig(
-    val context: String = "",
-    val keyPoints: List<String> = emptyList(),
-    val tone: String = "casual",
-)
-
-data class OnboardingContent(
-    val pages: List<OnboardingPage> = emptyList(),
-)
-
-data class OnboardingPage(
-    val title: String = "",
-    val description: String = "",
-    val emoji: String = "",
-)
-
 data class OnboardingAsset(
     val image: String,
 )
