+++ app/src/main/java/com/ilustris/sagai/features/onboarding/data/OnboardingStateMapper.kt
@@ -8,6 +8,7 @@ import com.ilustris.sagai.core.services.RemoteConfigService
 import com.ilustris.sagai.core.utils.StringResourceHelper
 import com.ilustris.sagai.features.home.data.model.Saga
 import com.ilustris.sagai.features.newsaga.data.model.Genre
+import com.ilustris.sagai.features.onboarding.data.model.OnboardingContent
 import com.ilustris.sagai.features.onboarding.ui.MorphingGenresBackground
 import com.ilustris.sagai.features.onboarding.ui.OnboardingAction
 import com.ilustris.sagai.features.onboarding.ui.OnboardingButton
