+++ app/src/main/java/com/ilustris/sagai/features/onboarding/domain/OnboardingUseCase.kt
@@ -2,8 +2,8 @@ package com.ilustris.sagai.features.onboarding.domain
 
 import com.ilustris.sagai.core.data.RequestResult
 import com.ilustris.sagai.features.newsaga.data.model.Genre
-import com.ilustris.sagai.features.onboarding.data.OnboardingContent
 import com.ilustris.sagai.features.onboarding.data.OnboardingType
+import com.ilustris.sagai.features.onboarding.data.model.OnboardingContent
 
 interface OnboardingUseCase {
     suspend fun shouldShow(type: OnboardingType): Boolean
