+++ gradle/gradle-daemon-jvm.properties
@@ -1,3 +1,3 @@
 #This file is generated by updateDaemonJvm
-toolchainVendor=jetbrains
-toolchainVersion=17
+toolchainVendor=openjdk
+toolchainVersion=21
